requirements.txt
celery==4.4.7
Django==3.1.2
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-timezone-field==4.0
djangorestframework==3.12.1
google-api-core==1.5.2
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
redis==3.5.3
requests==2.20.1
requests-oauthlib==1.0.0